* Honoring tags within config file. Closes #449
